backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
fhir.resources==8.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
python-multipart==0.0.9
